// src/types/aging.ts
export interface RemessaData {
  id?: string;
  numero_remessa: string; // Número da remessa (ex: 302604549)
  data_picking?: string; // Data de picking
  peso_total_remessa?: number; // Peso total da remessa
  item: string; // Número do item (ex: "10", "20")
  data_disponibilidade: string; // Data disponibilidade do material
  quantidade: number; // Quantidade da remessa
  unidade_medida: string; // Unidade (KG, UN, etc)
  material: string; // Código do material
  centro: string; // Centro (ex: 600)
  deposito: string; // Depósito (ALM, PES, PDD, etc)
  descricao_material: string; // Denominação do item
  created_at?: string;
  updated_at?: string;
}